Module cli_password_manager::encryption

Módulo que lida da encriptação e decriptação dos dados gerados

Visão Geral - AES

Advanced Encryption Standard (aes) também conhecida como cifra de Rijndael, é uma cifra de blocos de tamanho fixo simétrica, ou seja, divide a mensagem (texto que deseja-se cifrar) em blocos de N bits no qual realiza o processo de cifração/decifração, no qual ambos possuem mesma chave. N pode possuir os valores de 128, 194 ou 256 bits. Nas funções abaixo, a versão implementada da cifra de Rijndael foi a de 256 bits. Uma de suas vantagens diante aos algortimos de cifração que o sucedeu é o tamanho N bits que ataques de força bruta sejam inviáveis para a quebra de sigílo.

Blocos e rodadas

A mensagem é dividida em blocos de N bits e cada um desses blocos é aplicado o algoritmo de cifração e decifração da cifra de Rinjdael.

Tanto para cifrar tanto para decifrar, o aes executa o mesmo processo dividido em rodadas, diferem-se uma da outra apenas em relação a ordem dos processos executados. O número de rodadas a ser aplicado varia conforme o número de bits escolhido:

10 para 128 bits

12 para 194 bits

14 para 256 bits

Em cada uma delas, 4 métodos diferentes são aplicados sob o bloco a fim de produzir “maior aleatoriedade” no resultado final. O bloco possui um “estado” que é alterado a cada passo do algoritmo e é interpretado como uma matriz, no qual cada item da matriz refere-se a um byte.

Métodos

AddRoundKey

Cada byte do estado é combinado com um byte da chave de rodada utilizando da operação XOR bit a bit.

SubBytes

Substituição não linear dos bytes utilizando-se de uma tabela de pesquisa, por padrão a S-Box

ShiftRows

Opera nas linhas do estado, fazendo permutação dos bytes de uma mesma linha

MixColumns

Fornece difusão na cifra, no qual cada coluna da matriz é transformada utilizando-se de uma matriz fixa. As entradas são tratadas como polinômios sobre GF (2⁸).

Cifração e decifração por rodada

O processo de cifração ocorre na seguinte ordem:

SubBytes

ShiftRows

MixColumns

AddRoundKey Com exceção do último round que não faz a operação de mix columns

O processo de decifração ocorre na seguinte ordem:

Inverse ShiftRows

Inverse SubBytes

AddRoundKey

Inverse MixColumns Com exceção do último round que não faz a operação de inverse mix columns

Functions

decrypt_data

Função de decifrar

encrypt_data

Função de encriptar

generate_key

Função para gerar hash de 256 bits