Function cli_password_manager::encryption::encrypt_data

pub fn encrypt_data(data: &[u8], key: &[u8]) -> Result<Vec<u8>, ErrorStack>

Função de encriptar

Parâmetros

data: toda a mensagem que deseja-se cifrar, já no formato de bytes

key: chave de 256 bits que será usada

Implementação

Chama o construtor cipher da biblioteca openssl para termos uma instância da cifra de Rinjdael de 256 bits A variável iv é o vetor de inicialização aleatório de 256 bits que faz xor bit a bit no processo inicial de cifração a fim de aumentar a aleatoriedade da cifração. Por fim chamamos a função encrypt da biblioteca openssl que faz todo o processo descrito anteriormente.

Retorno

Retornamos o texto cifrado utilizando a cifra de blocos aes 256 bits.